#316 [refactor] time 패키지 구조 변경
