+++ src/main/java/com/asap/server/service/UserService.java
@@ -20,8 +20,8 @@ import com.asap.server.persistence.repository.user.UserRepository;
 import com.asap.server.presentation.controller.dto.request.UserRequestDto;
 import com.asap.server.presentation.controller.dto.response.UserMeetingTimeResponseDto;
 import com.asap.server.presentation.controller.dto.response.UserTimeResponseDto;
-import com.asap.server.service.dto.UserMeetingScheduleRegisterDto;
-import com.asap.server.service.dto.UserTimeRegisterDto;
+import com.asap.server.service.time.dto.UserMeetingScheduleRegisterDto;
+import com.asap.server.service.time.dto.UserTimeRegisterDto;
 import com.asap.server.service.time.UserMeetingScheduleService;
 import com.asap.server.service.vo.BestMeetingTimeVo;
 import com.asap.server.service.vo.BestMeetingTimeWithUsersVo;
