+++ src/main/java/com/asap/server/service/time/dto/UserTimeRegisterDto.java
@@ -1,4 +1,4 @@
-package com.asap.server.service.dto;
+package com.asap.server.service.time.dto;
 
 import java.util.List;
